requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloud-tpu-client==0.10
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docutils==0.21.2
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h5py==3.11.0
httplib2==0.22.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
KivyOnTop==1.4
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mouse==0.7.1
MouseInfo==0.1.3
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==3.0.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.0.3
